7/03/2024-RETURNING QUOTES WRT TO LIMIT

+++ src/main/java/com/example/quotely/demo/Service/QuoteService.java
@@ -4,7 +4,7 @@ import com.example.quotely.demo.Vo.QuotesVo;
 import com.example.quotely.demo.Vo.ResponseData;
 
 public interface QuoteService {
-    public ResponseData newQuote(Integer limit, Long userId);
+    public ResponseData newQuote(Long limit, Long userId);
 
    public ResponseData addQuote(QuotesVo quotesVo);
 }
